Change the position of time duration
