+++ src/components/AddressDetail/AddressHeader/index.tsx
@@ -15,8 +15,9 @@ import BookmarkButton from "src/components/commons/BookmarkIcon";
 import TokenAutocomplete from "src/components/TokenAutocomplete";
 import ADAicon from "src/components/commons/ADAIcon";
 import { useScreen } from "src/commons/hooks/useScreen";
+import FormNowMessage from "src/components/commons/FormNowMessage";
 
-import { BackButton, BackText, RedirectButton, StyledBoxCard, TitleText, WrapHeader } from "./styles";
+import { BackButton, BackText, RedirectButton, StyledBoxCard, TimeDuration, TitleText, WrapHeader } from "./styles";
 
 interface Props {
   data: WalletAddress | null;
@@ -24,9 +25,11 @@ interface Props {
 }
 const AddressHeader: React.FC<Props> = ({ data, loading }) => {
   const [stakeKey, setStakeKey] = useState("");
-  const { data: dataStake, loading: loadingStake, lastUpdated } = useFetch<WalletStake>(
-    stakeKey ? `${API.STAKE.DETAIL}/${stakeKey}` : ""
-  );
+  const {
+    data: dataStake,
+    loading: loadingStake,
+    lastUpdated
+  } = useFetch<WalletStake>(stakeKey ? `${API.STAKE.DETAIL}/${stakeKey}` : "");
   const { adaRate } = useSelector(({ system }: RootState) => system);
   const theme = useTheme();
   const { isMobile } = useScreen();
@@ -98,7 +101,6 @@ const AddressHeader: React.FC<Props> = ({ data, loading }) => {
         <Box
           width={"100%"}
           display={"flex"}
-          pb={2}
           flexWrap={"wrap"}
           alignItems={"center"}
           justifyContent={"space-between"}
@@ -117,6 +119,9 @@ const AddressHeader: React.FC<Props> = ({ data, loading }) => {
             </RedirectButton>
           )}
         </Box>
+        <TimeDuration>
+          <FormNowMessage time={lastUpdated} />
+        </TimeDuration>
       </WrapHeader>
       <Grid container spacing={2}>
         <Grid item xs={12} md={6}>
@@ -127,7 +132,6 @@ const AddressHeader: React.FC<Props> = ({ data, loading }) => {
               address={data?.address || ""}
               item={itemLeft}
               loading={loading}
-              lastUpdated={lastUpdated}
             />
           </StyledBoxCard>
         </Grid>
