+++ src/components/share/CardAddress/index.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { Box, Skeleton } from "@mui/material";
 
-import FormNowMessage from "src/components/commons/FormNowMessage";
 import { EmptyIcon, InfoIcon } from "src/commons/resources";
 import CopyButton from "src/components/commons/CopyButton";
 import CustomIcon from "src/components/commons/CustomIcon";
@@ -14,7 +13,6 @@ import {
   ItemDetail,
   LabelItem,
   RowItem,
-  TimeDuration,
   TitleDetail,
   TokenAddress,
   ValueItem
@@ -27,9 +25,8 @@ interface DetailCardProps {
   type: "left" | "right";
   loading: boolean;
   addressDestination?: string;
-  lastUpdated?: number;
 }
-const CardAddress: React.FC<DetailCardProps> = ({ title, address, item, type, loading, addressDestination, lastUpdated }) => {
+const CardAddress: React.FC<DetailCardProps> = ({ title, address, item, type, loading, addressDestination }) => {
   if (loading) {
     return (
       <CardItemStyled padding={0}>
@@ -50,11 +47,6 @@ const CardAddress: React.FC<DetailCardProps> = ({ title, address, item, type, lo
   return (
     <CardItemStyled>
       <TitleDetail>{title}</TitleDetail>
-      {type === "left" ? (
-        <TimeDuration>
-          <FormNowMessage time={lastUpdated} />
-        </TimeDuration>
-      ) : null}
       <AddressGroup>
         {type === "left" ? (
           <TokenAddress>{address}</TokenAddress>
