+++ src/components/share/CardAddress/styles.ts
@@ -82,11 +82,3 @@ export const CardItemStyled = styled(CardItem)`
     padding: 20px 15px;
   }
 `;
-
-export const TimeDuration = styled("small")(({ theme }) => ({
-  color: theme.palette.grey[500],
-  display: "block",
-  textAlign: "left",
-  flex: 1,
-  paddingTop: "10px"
-}));
